Refactor: Move software texture mapping to dedicated file
